requirements.txt
fastapi
langchain
pypdf
google-genai
python-dotenv
sentence-transformers
faiss-cpu
